pyproject.toml
[project]
name = "cv-analyzer-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.4.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "pytest>=8.4.1",
    "python-decouple>=3.8",
    "supabase>=2.16.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
